fix: LottoEvent 제거
getter와 setter만 가지고있는 LottoEvent를 LottoController로 옮기고 제거
https://github.com/next-step/java-lotto/pull/2024#discussion_r698306447
